Add ship cache invalidation and move tests around
